web scraping agent/requirements.txt
# Core
langchain
langchain-core
python-dotenv
pydantic
requests
beautifulsoup4
lxml

# LLM backends (install the ones you use)
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-google-vertexai

# Vertex AI often needs Application Default Credentials + project; optional:
# google-cloud-aiplatform
